requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-login==0.6.2
flask-wtf==1.2.1
flask-bcrypt==1.0.1
python-dotenv==1.0.0
plaid-python==16.2.0
cryptography==41.0.4
werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
wtforms==3.0.1
jinja2==3.1.2
itsdangerous==2.1.2
pytest==7.4.2
pytest-flask==1.2.0
